feat: 添加缓冲区功能、相关接口

+++ server/api.js
@@ -549,16 +549,41 @@ const readOriginGeo = async (request, response) => {
   response.status(200).json({ status: true, results: originGeoJSONArr })
 }
 
-//  删除图层 同时删除原来geojson文件
+//  删除图层 同时删除 数据库中的geojson表 有文件的删除geojson文件
 const delLayers = async (request, response) => {
   const layer = request.body.layer
   const layerLower = layer.toLowerCase()
   try {
-    fs.unlinkSync('./public/' + layer + '.geojson')
+    let allFiles = readFileList('./public')
+    allFiles.forEach(filepath => {
+      const file = filepath.split('.')[0].split('\\')[1]
+      if (file === layer) fs.unlinkSync('./public/' + layer + '.geojson')
+    })
     await pool.query(`DROP TABLE ${layerLower};`)
     response.status(200).json({ status: true, message: '图层删除成功' })
   } catch {
     response.status(500).json({ status: false, message: '图层删除失败' })
+  }
+}
+
+//  生成geojson 上传到数据库
+const generateGeoJson = async (request, response) => {
+  const newGeoJSON = request.body.newGeoJSON
+  const radius = request.body.radius
+  const tablename = 'buffer' + radius
+  try {
+    const q1 = `select count(*) from pg_class where relname = '${tablename}';`
+    let res = await pool.query(q1)
+    if (res.rows[0].count !== '0') {
+      await pool.query(`DROP TABLE ${tablename};`)
+    }
+    await insertData(tablename, newGeoJSON.features)
+    let layerData = { layer: tablename, layerVisual: true }
+    response.status(200).json({ status: true, result: newGeoJSON, layerData: layerData })
+  } catch (err) {
+    console.log(err)
+    await pool.query(`DROP TABLE buffer'${radius};`)
+    response.status(500).json({ status: false, message: '数据出错，未存到数据库中' })
   }
 }
 module.exports = {
@@ -584,5 +609,6 @@ module.exports = {
   uploadDatabase,
   layerProperty,
   readOriginGeo,
-  delLayers
+  delLayers,
+  generateGeoJson
 }
